% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logentries.r
\name{le_acct_key}
\alias{le_acct_key}
\title{Get or set LOGENTRIES_ACCOUNT_KEY value}
\usage{
le_acct_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new LogEntries account API key for the current environment?}
}
\value{
atomic character vector containing the LOGENTRIES_ACCOUNT_KEY API key
}
\description{
The API wrapper functions in this package all rely on a LogEntries account API
key residing in the environment variable \code{LOGENTRIES_ACCOUNT_KEY}. The
easiest way to accomplish this is to set it in the `.Renviron` file in your
home directory.
}

