% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logentries.r
\name{le_read_log}
\alias{le_read_log}
\title{Read log data from LogEntries}
\usage{
le_read_log(log_set = "DemoSet", log_name = "Usage_trail", start = NULL,
  end = NULL, filter = NULL, limit = NULL, acct_key = le_acct_key())
}
\arguments{
\item{start, end}{start and end timestamps (ms from epoch)}

\item{filter}{keyword filter}

\item{limit}{how many entries to return?}

\item{acct_key}{LogEntries account key}

\item{logs_set, log_name}{names of the set & log}
}
\description{
This is the simple interface to LogEntries
}

