% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logentries.r
\name{le_write_log}
\alias{le_write_log}
\title{Post log data to LogEntries}
\usage{
le_write_log(log_data, log_token)
}
\arguments{
\item{log_data}{the text to log. No conversion is done but LogEntries expectes
the content to be textual and 8192 characters or less.}

\item{log_token}{api token for the log}
}
\description{
You need to setup a \href{https://docs.logentries.com/docs/input-token}{log token}
which should be supplied as the \code{log_token} parameter.
}

